+++ .jshintignore
@@ -1,0 +1,1 @@
+node_modules/
