+++ .npmignore
@@ -1,0 +1,4 @@
+node_modules/
+test/
+.gitignore
+.editorconfig
